skills/ai-news-crawler/requirements.txt
feedparser
arxiv
beautifulsoup4
lxml
requests
jinja2
pyyaml
